string constants added to messages.py
